Game server connection event, scene creator, and better logging
Added a way for the game server to send errors during connection.
Added a log function to the servers.
Refactored the way users that need be kicked are handled.

Fixed isort

The game server now gets the players as argument

Removed useless argument to flake8 in pong-server CI linter test

Added a wait time before users are kicked to make sure the clients have enough time to disconnect on their own

Made it so when a user connects twice (to the game server) the first connected client will be disconnected and replaced by the second

The game scene is now sent to the client once all the players have joined

Updated documentation

The game server logs are now displayed by the redirection server when debug is enabled

Replaced double quotes with simple quotes

Now uses logging module

Added some tests for the redirection server